Added firebase authentication screen
