+++ lib/view/auth_gate/auth_gate.dart
@@ -1,0 +1,22 @@
+import 'package:excercise/main.dart';
+import 'package:excercise/view/screens/my_home_page.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_ui_auth/firebase_ui_auth.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class AuthGate extends StatelessWidget {
+  const AuthGate({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return StreamBuilder(
+        stream: FirebaseAuth.instance.userChanges(),
+        builder: (context, snapshot) {
+          if (!snapshot.hasData) {
+            return const MyApp(child: SignInScreen());
+          }
+          return const MyApp(child: ProviderScope(child: MyHomePage()));
+        });
+  }
+}
