+++ pubspec.yaml
@@ -43,6 +43,9 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   intl: any
+  firebase_core: ^2.28.0
+  firebase_auth: ^4.19.0
+  firebase_ui_auth: ^1.13.1
 
 dev_dependencies:
   flutter_test:
