+++ test/widget_test.dart
@@ -5,7 +5,9 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
+import 'package:excercise/view/screens/my_home_page.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:excercise/main.dart';
@@ -13,7 +15,8 @@ import 'package:excercise/main.dart';
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester
+        .pumpWidget(const MyApp(child: ProviderScope(child: MyHomePage())));
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
